MG468 - Enhanced join support (MapGuide RFC 5)
git-svn-id: http://svn.osgeo.org/mapguide/trunk@1020 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Server/src/Gws/GwsCommon/inc/GwsFeatureIdSet.h
@@ -54,7 +54,7 @@ public:
             return false;
         mVecIds.push_back(id);
         count++;
-        assert(count == (int)mVecIds.size());
+        assert(count == mVecIds.size());
         return true;
     }
 
@@ -81,7 +81,7 @@ public:
         }
 
         count--;
-        assert(count == (int)mVecIds.size());
+        assert(count == mVecIds.size());
         return true;
     }
 
@@ -117,4 +117,4 @@ private:
     FdoInt32                    count;
 };
 
-#endif
+#endif /* GwsFeatureIdSet_h */
